requirements.txt
pdfminer.six==20221105
requests==2.31.0